[sim] fixed multiply-high in rv32

+++ riscv/insns/mulhu.h
@@ -1,4 +1,4 @@
 if(xpr64)
   RD = (uint128_t(RS1) * uint128_t(RS2)) >> 64;
 else
-  RD = sext32((RS1 * RS2) >> 32);
+  RD = sext32(((uint64_t)(uint32_t)RS1 * (uint64_t)(uint32_t)RS2) >> 32);
